Move components into the common/ subfolder Move the login and signup components into the common/ subfolder Move the header and footer components into the common/ subfolder Move the Search component into the common/ subfolder Modify paths to these components in other components that import them
