+++ client/components/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Footer from './Footer';
-import Header from './Header';
+import Footer from './common/Footer';
+import Header from './common/Header';
 import Main from './Main';
 
 const App = () => (
