+++ client/components/centers/containers/Centers.jsx
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import * as centerActions from '../../../actions/centerActions';
 import CenterList from '../presentational/CenterList';
-import Search from '../../Search';
+import Search from '../../common/Search';
 import Preloader from '../../common/Preloader';
 
 class Centers extends Component {
